(refactor) components folder structure
