+++ src/components/modals/TransferOutcomeTokensModal/index.tsx
@@ -4,12 +4,12 @@ import styled from 'styled-components'
 
 import { Button } from 'components/buttons'
 import { Modal, ModalProps } from 'components/common/Modal'
-import { StatusInfoInline, StatusInfoType } from 'components/common/StatusInfoInline'
 import { Amount } from 'components/form/Amount'
 import { InputAddress } from 'components/form/InputAddress'
 import { ButtonContainer } from 'components/pureStyledComponents/ButtonContainer'
 import { Row } from 'components/pureStyledComponents/Row'
 import { InlineLoading } from 'components/statusInfo/InlineLoading'
+import { StatusInfoInline, StatusInfoType } from 'components/statusInfo/StatusInfoInline'
 import { ZERO_BN } from 'config/constants'
 import { useWeb3Connected } from 'contexts/Web3Context'
 import { ERC20Service } from 'services/erc20'
