+++ src/pages/MergePositions/Contents.tsx
@@ -4,7 +4,6 @@ import React, { useCallback, useEffect, useMemo, useState } from 'react'
 
 import { Button } from 'components/buttons/Button'
 import { CenteredCard } from 'components/common/CenteredCard'
-import { StatusInfoInline, StatusInfoType } from 'components/common/StatusInfoInline'
 import { Amount } from 'components/form/Amount'
 import { SelectCondition } from 'components/form/SelectCondition'
 import { SelectPositions } from 'components/form/SelectPositions'
@@ -13,6 +12,7 @@ import { MergeResultModal } from 'components/mergePositions/MergeResultModal'
 import { ButtonContainer } from 'components/pureStyledComponents/ButtonContainer'
 import { Row } from 'components/pureStyledComponents/Row'
 import { FullLoading } from 'components/statusInfo/FullLoading'
+import { StatusInfoInline, StatusInfoType } from 'components/statusInfo/StatusInfoInline'
 import { IconTypes } from 'components/statusInfo/common'
 import { NULL_PARENT_ID, ZERO_BN } from 'config/constants'
 import { useBatchBalanceContext } from 'contexts/BatchBalanceContext'
