+++ src/pages/PrepareCondition/index.tsx
@@ -4,11 +4,11 @@ import { useHistory } from 'react-router-dom'
 
 import { Button } from 'components/buttons/Button'
 import { ButtonType } from 'components/buttons/buttonStylingTypes'
-import { ArbitratorDropdown } from 'components/common/ArbitratorDropdown'
-import { CategoriesDropdown } from 'components/common/CategoriesDropdown'
 import { CenteredCard } from 'components/common/CenteredCard'
-import { ConditionTypesDropdown } from 'components/common/ConditionTypesDropdown'
 import { AddOutcome } from 'components/form/AddOutcome'
+import { ArbitratorDropdown } from 'components/form/ArbitratorDropdown'
+import { CategoriesDropdown } from 'components/form/CategoriesDropdown'
+import { ConditionTypesDropdown } from 'components/form/ConditionTypesDropdown'
 import { ButtonContainer } from 'components/pureStyledComponents/ButtonContainer'
 import { ErrorContainer, Error as ErrorMessage } from 'components/pureStyledComponents/Error'
 import { PageTitle } from 'components/pureStyledComponents/PageTitle'
